Increment 5D2: preserve truthful non-complete authority outcomes
